+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## vNext
 
+- Support querying by date range for CAISO, PJM, NYISO, and ISONE Hisotrical Methods
 - Add gas prices to CAISO
+- Add GHG allowance price to CAISO
 
 ## v0.8.0 - Oct 13, 2022